refactor(4a): extract attach fan-out + gate promotion into daemon/attach.ts
Move broadcast, schedulePromotion, and cancelPromotionsFor into
createAttach(ctx). cockpitd.ts calls createAttach and assigns the
returned handlers back onto ctx so the driver emit closures can
reference them lazily via the context object.
