+++ src/control/cockpitd.ts
@@ -2,9 +2,9 @@
 import { join, dirname } from "node:path";
 import { fileURLToPath } from "node:url";
 import { readFileSync } from "node:fs";
-import { randomUUID } from "node:crypto";
 import { createDaemon } from "./daemon.js";
 import { buildContext } from "./daemon/context.js";
+import { createAttach } from "./daemon/attach.js";
 export type { CockpitdOpts } from "./daemon/context.js";
 export { defaultIsPidAlive } from "./daemon/context.js";
 import { startServer, encodeFrame, type AttachFrame, type AttachInbound } from "./protocol.js";
@@ -12,7 +12,6 @@ import { runHeadless } from "./headless-launcher.js";
 import { CodexInteractiveDriver, shouldReattachCodex } from "./codex/driver.js";
 import { OpencodeSseBridge } from "./opencode/sse-bridge.js";
 import { CmuxEventsBridge } from "./cmux/events-bridge.js";
-import { makeGate } from "./codex/gate.js";
 import { appendToMailbox, rotateIfNeeded, mailboxStats, readCursor, writeCursor, readFromCursor } from "./mailbox.js";
 import { createRelayHealer } from "./relay-healer.js";
 import { createDirectSurfaceLivenessProbe, createDirectCrewPaneReader } from "./crew-pane-reader.js";
@@ -22,7 +21,7 @@ import { projectHealth, type ComponentHealth } from "./liveness.js";
 import { assembleDaemonSnapshot, type DaemonSnapshotInputs } from "./snapshot.js";
 import { loadConfig } from "@cockpit/shared";
 import { readdir } from "node:fs/promises";
-import type { Gate, TaskRecord, ControlEvent } from "@cockpit/shared";
+import type { TaskRecord, ControlEvent } from "@cockpit/shared";
 import { TERMINAL_STATES } from "@cockpit/shared";
 import type { Socket } from "node:net";
 import type { PaneRef } from "../runtimes/types.js";
@@ -80,58 +79,10 @@ export function startCockpitd(opts: CockpitdOpts = {}) {
     return probeResults.get(rec.id) ?? "unknown";
   };
 
-  function broadcast(taskId: string, f: AttachFrame): void {
-    const conns = attachConns.get(taskId);
-    if (!conns) return;
-    const wire = encodeFrame(f);
-    for (const conn of conns) {
-      try { conn.write(wire); } catch { /* client gone; onAttachClose will clean up */ }
-    }
-  }
-
-  // ── Gate promotion (spec §4.9) ────────────────────────────────────────────
-  // When a server-request event fires and no client is attached, start a 5s
-  // timer. If still unattached at fire time, promote to a Gate in the store
-  // and broadcast gate-promoted so any later-attaching client can offer takeover.
-  const pendingGateTimers = new Map<string, { taskId: string; timer: NodeJS.Timeout }>();
-
-  function schedulePromotion(
-    taskId: string,
-    requestId: number,
-    kind: "input" | "approval",
-    question: string,
-  ): void {
-    // If a client is already attached for this task, no promotion needed.
-    const conns = attachConns.get(taskId);
-    if (conns && conns.size > 0) return;
-    const key = `${taskId}#${requestId}`;
-    // Clear any prior timer for the same (taskId, requestId).
-    const prior = pendingGateTimers.get(key);
-    if (prior) clearTimeout(prior.timer);
-    const timer = setTimeout(() => {
-      pendingGateTimers.delete(key);
-      // Re-check at fire time — a client may have attached in the 5s window.
-      if (attachConns.get(taskId)?.size) return;
-      const rec = store.listAll().find((r) => r.id === taskId);
-      if (!rec) return;
-      const gate: Gate = makeGate({ taskId, kind, question, now: Date.now(), mkId: () => randomUUID() });
-      const gates = [...(rec.gates ?? []), gate];
-      store.put({ ...rec, gates });
-      broadcast(taskId, { type: "gate-promoted", taskId, gateId: gate.gateId });
-      log(`gate promoted gateId=${gate.gateId} taskId=${taskId} kind=${kind}`);
-    }, 5_000);
-    timer.unref?.();
-    pendingGateTimers.set(key, { taskId, timer });
-  }
-
-  function cancelPromotionsFor(taskId: string): void {
-    for (const [key, slot] of pendingGateTimers.entries()) {
-      if (slot.taskId === taskId) {
-        clearTimeout(slot.timer);
-        pendingGateTimers.delete(key);
-      }
-    }
-  }
+  const { broadcast, schedulePromotion, cancelPromotionsFor } = createAttach(ctx);
+  ctx.broadcast = broadcast;
+  ctx.schedulePromotion = schedulePromotion;
+  ctx.cancelPromotionsFor = cancelPromotionsFor;
 
   // ── CodexInteractiveDriver singleton ─────────────────────────────────────
   // The driver holds the single AppServerClient child. Each task maps to a
